+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
+[v6.1.0-13] (Unreleased)
+Bug fixes & Enhancements:
+[#272] (https://github.com/HewlettPackard/terraform-provider-oneview/issues/272) more server profile/template issues not working as expected.
+[#295] (https://github.com/HewlettPackard/terraform-provider-oneview/issues/295) server profile update in-place.
+[#288] (https://github.com/HewlettPackard/terraform-provider-oneview/issues/288) Error updating logical interconnect group Synergy OneView 5.5
+
 # [v6.0.0-13]
 #### Major changes
 1.This release is validated to support Terraform 0.13.  Terraform 0.13 support HCL2 and enables automated installation of the HPE OneView Terraform Provider.